pyproject.toml
[project]
name = "spectra-pipe"
version = "0.1.0"
description = "A Hyperspectral Image Processing CLI Utility!"
authors = [
    {name = "fernandev",email = "fernando.daza.18@outlook.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12.7,<3.15"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "scikit-image (>=0.26.0,<0.27.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "numpy (==2.2.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "scipy (>=1.16.3,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "h5py (>=3.15.1,<4.0.0)",
    "torch (>=2.9.1,<3.0.0)",
    "torchvision (>=0.24.1,<0.25.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "typer (>=0.9.0,<1.0.0)",
    "pyyaml (>=6.0,<7.0)",
    "einops (>=0.7.0,<1.0.0)",
    "rich (>=13.0.0,<14.0.0)",
]

[project.scripts]
spectrapipe = "hsi_pipeline.cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "hsi_pipeline", from = "src"}]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "slow: marks tests as slow (run with '--slow')",
]

[dependency-groups]
dev = [
    "pytest (>=8.3.3,<9.0.0)",
    "pytest-cov (>=6.0.0,<7.0.0)",
    "ruff (>=0.13.0,<0.14.0)",
]

[tool.ruff]
exclude = ["external"]
line-length = 88
